Issue #839 Refactor to move service HSDS data validation to the new validation folder
