+++ src/stores/services/index.ts
@@ -1,9 +1,9 @@
 import * as R from 'ramda';
 import * as constants from '../../application/constants';
 import * as actions from './actions';
-import * as types from './types';
+import * as types from '../../validation/services/types';
 
-import { Id, ServiceStore } from './types';
+import { Id, ServiceStore } from '../../validation/services/types';
 export { Id, ServiceStore };
 
 export function buildDefaultStore(): types.ServiceStore {
