+++ src/stores/services/is_error_services_for_topic.ts
@@ -1,4 +1,4 @@
-import { ErrorServicesForTopic, ServicesForTopic } from './types';
+import { ErrorServicesForTopic, ServicesForTopic } from '../../validation/services/types';
 import * as constants from '../../application/constants';
 
 export const isErrorServicesForTopic = (services: ServicesForTopic):
