+++ src/validation/services/index.ts
@@ -4,7 +4,7 @@ const Ajv = require('ajv');
 
 import * as R from 'ramda';
 import * as types from './types';
-import { serviceAtLocationArray } from './json_schema';
+import { serviceAtLocationArray } from './schema';
 
 // tslint:disable-next-line:no-any
 export const validateServicesAtLocationArray = (data: any): ValidationResult => {
